requirements.txt
flask>=2.0
tensorflow>=2.11
Pillow>=9.0
numpy>=1.22
requests>=2.28
